Updates urllib3 version and tox adjustment
An issue was found via dependabot alerting us the
urllib3 version need to be bumped.  When changing
version, it required the requests library version
to change as well for compatibility.

This patch updates the tox file [testenv] to provide
additional Python versions and adds code coverage found
in engine and enterprise.

Fixes #ENTERPRISE-562

Signed-off-by: Ryan Brady <ryan.brady@anchore.com>