Fixed Functional Tests for Anchore CLI
The tests themselves were not broken, per se, but rather a new
vulnerability was found in the test image that is being used in the
tests, specifically in NumPy, that was discovered in mid-December 2021.
This happens to coincide with when our CLI functional tests started to
fail.

This patch updates the tests to expect the correct number of
vulnerabilities and add the new NumPy vulnerability data to other tests
to make sure that we're checking for that information as well.

I will note, we should probably use a different image from the one we're
using currently.

Signed-off-by: Dustin Schoenbrun <dustin.schoenbrun@anchore.com>